requirements.txt
# Core dependencies
numpy<1.24.0
chumpy>=0.70
opencv-python>=4.5.0
pillow>=8.0.0
scipy>=1.7.0
tqdm>=4.62.0
matplotlib>=3.4.0

# Data handling
omegaconf>=2.1.0
h5py>=3.1.0
lpips>=0.1.4

# Deep learning frameworks
torch>=1.10.0
torchvision>=0.11.0

# 3D and mesh processing
trimesh>=3.9.0
pyrender>=0.1.45
smplx>=0.1.28
chumpy>=0.70
scikit-image>=0.18.0